+++ src/boxrec-pages/profile/boxrec.page.profile.ts
@@ -86,6 +86,10 @@ export abstract class BoxrecPageProfile extends BoxrecParseBoutsParseBouts {
         return this.parseOtherInfo();
     }
 
+    get picture(): string {
+        return this.$(".profileTablePhoto img").attr("src");
+    }
+
     /**
      * Returns the current residency of the person
      * @returns {Location}